requirements.txt
# https://github.com/halcy/Mastodon.py
mastodon.py
